refactor TxnStatusChangeReason
